refactor: DeadlineTimelineコンポーネントを分離
- 300行のIIFEを独立コンポーネントに切り出し
- ポモドーロ開始ロジックをstartPomodoroWithTaskヘルパーに集約
- 未使用の@google/genai依存を削除
- useMemoでタスクフィルタリングを最適化
